+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Version 1.6
 - Verify element state after click, for sites with persistent toggle
+- Add logging for callback errors
 
 ## Version 1.5.1
 - Additional YouTube selector